Updates following suggestions from PR discussion.
Among others:
-  Moved TrainState into flax.training.train_state.
- Also renamed TrainState.update() to TrainState.apply_gradients().
- Using optax.sgd().
