+++ examples/imagenet/train_test.py
@@ -44,8 +44,8 @@ class TrainTest(absltest.TestCase):
   def test_create_model(self):
     """Tests creating model."""
     model = train.create_model(model_cls=models._ResNet1, half_precision=False)  # pylint: disable=protected-access
-    params, state = train.initialized(random.PRNGKey(0), 224, model)
-    variables = {'params': params, **state}
+    params, batch_stats = train.initialized(random.PRNGKey(0), 224, model)
+    variables = {'params': params, 'batch_stats': batch_stats}
     x = random.normal(random.PRNGKey(1), (8, 224, 224, 3))
     y = model.apply(variables, x, train=False)
     self.assertEqual(y.shape, (8, 1000))
